翻译
In 2003 Google published an inconspicuous paper titled “The Google Filesystem” ( http://static.googleusercontent.com/media/research.google.com/en/us/archive/gfs-sosp2003.pdf ). Not many outside of Silicon Valley paid much attention to its publication or the message it was trying to convey. The message it told was directly applicable to a company like Google, whose primary business focused on indexing the Internet, which was not a common use case for most companies. The paper described a storage framework uniquely designed to handling the current future technological demands Google envisioned for its business. In the spirit of TL&DR, here are its most salient points:

- Failures are the norm
- Files are large
- Files are changed by appending, not by updating
- Closely coupled application and filesystem APIs